Add DAO spec
